Update MM tests
