+++ test/unit_tests/js/organisms/MegaMenu-spec.js
@@ -1,2093 +1,6 @@
 import MegaMenu from '../../../../cfgov/unprocessed/js/organisms/MegaMenu';
 import { simulateEvent } from '../../../util/simulate-event';
-
-const BASE_CLASS = 'o-mega-menu';
-const HTML_SNIPPET = `
-<nav class="o-mega-menu
-            u-hidden" data-js-hook="behavior_flyout-menu" aria-label="main menu">
-    <button class="o-mega-menu_trigger" data-js-hook="behavior_flyout-menu_trigger" aria-haspopup="menu">
-        <span class="o-mega-menu_trigger-open">
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 784.5 1200" class="cf-icon-svg"><path d="M65 370.5h654.5c35.9 0 65-29.1 65-65s-29.1-65-65-65H65c-35.9 0-65 29.1-65 65s29.1 65 65 65zM719.5 515H65c-35.9 0-65 29.1-65 65s29.1 65 65 65h654.5c35.9 0 65-29.1 65-65s-29.1-65-65-65zM719.5 789.4H65c-35.9 0-65 29.1-65 65s29.1 65 65 65h654.5c35.9 0 65-29.1 65-65s-29.1-65-65-65z"></path></svg>
-        </span>
-        <span class="o-mega-menu_trigger-close">
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 718.9 1200" class="cf-icon-svg"><path d="M451.4 613.7l248.1-248.1c25.6-25.1 26-66.3.8-91.9s-66.3-26-91.9-.8l-.8.8-248.1 248.1-248.1-248.1c-25.4-25.4-66.5-25.4-91.9 0s-25.4 66.5 0 91.9l248.1 248.1L19.5 861.8c-25.6 25.1-26 66.3-.8 91.9s66.3 26 91.9.8l.8-.8 248.1-248.1 248.1 248.1c25.4 25.4 66.5 25.4 91.9 0s25.4-66.5 0-91.9L451.4 613.7z"></path></svg>
-        </span>
-        <span class="u-visually-hidden">
-            Main menu
-        </span>
-    </button>
-
-
-<section class="o-mega-menu_content o-mega-menu_content-1 " aria-expanded="false" data-js-hook="behavior_flyout-menu_content">
-
-
-
-
-    <div class="o-mega-menu_content-wrapper o-mega-menu_content-1-wrapper ">
-
-
-
-
-
-
-
-        <div class="o-mega-menu_content-lists o-mega-menu_content-1-lists ">
-
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-1-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-1-list ">
-
-
-            <ul>
-
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-1-item o-mega-menu_content-item__has-icon o-mega-menu_content-1-item__has-icon ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-1-link " href="/complaint/">
-<h2>
-<span class="o-mega-menu_content-link-icon-pre o-mega-menu_content-1-link-icon-pre "><svg class="cf-icon-svg" viewBox="0 0 825.1 1200" xmlns="http://www.w3.org/2000/svg"><path d="M755.1 263.9H70c-38.6.1-69.9 31.4-70 70v459.5c.1 38.6 31.4 69.9 70 70h268.4l56.9 91.2c8.7 14 23 14 31.8 0l56.9-91.2h271.1c38.6-.1 69.9-31.4 70-70V333.9c-.1-38.7-31.4-69.9-70-70zm-375 132.5c0-17.9 14.6-32.5 32.5-32.5s32.5 14.6 32.5 32.5v204.2c0 17.9-14.6 32.5-32.5 32.5s-32.5-14.6-32.5-32.5V396.4zm32.5 364.4c-24.6 0-44.6-20-44.6-44.6s20-44.6 44.6-44.6 44.6 20 44.6 44.6c-.1 24.6-20 44.6-44.6 44.6z"></path></svg></span>
-
-
-        Submit a Complaint
-
-
-
-
-        </h2>
-</a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-1-item " data-js-hook="behavior_flyout-menu">
-
-    <a aria-haspopup="menu" class="o-mega-menu_content-link o-mega-menu_content-1-link o-mega-menu_content-link__has-children o-mega-menu_content-1-link__has-children " data-js-hook="behavior_flyout-menu_trigger" href="/consumer-tools/">
-<h2>
-
-
-
-
-        Consumer Tools
-
-
-        <span class="o-mega-menu_content-link-icon-post o-mega-menu_content-1-link-icon-post "><svg class="cf-icon-svg" viewBox="0 0 559.6 1200" xmlns="http://www.w3.org/2000/svg"><path d="M65.1 1090.2c-35.9 0-65-29-65.1-64.9 0-17.3 6.8-33.9 19.1-46.1l383.6-383.5L19.1 212.2c-25.1-25.6-24.8-66.8.9-92 25.3-24.8 65.8-24.8 91.1 0l429.5 429.5c25.4 25.4 25.4 66.5 0 91.9L111 1071.2c-12.1 12.2-28.7 19.1-45.9 19z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-closed o-mega-menu_content-1-link-icon-closed "><svg class="cf-icon-svg" viewBox="0 0 988.2 1200" xmlns="http://www.w3.org/2000/svg"><path d="M494.1 967.2c-17.3 0-33.8-6.8-46-19L18.6 518.6c-25.1-25.6-24.8-66.8.8-91.9 25.3-24.8 65.8-24.8 91.1 0l383.6 383.6 383.6-383.6c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L540.1 948.1c-12.2 12.2-28.7 19.1-46 19.1z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-open o-mega-menu_content-1-link-icon-open "><svg class="cf-icon-svg" viewBox="0 0 988.7 1200" xmlns="http://www.w3.org/2000/svg"><path d="M923.6 967.6c-17.3 0-33.8-6.8-46-19L494.1 565 110.5 948.5c-25.6 25.1-66.8 24.8-91.9-.8-24.8-25.3-24.8-65.8 0-91.1l429.5-429.5c25.4-25.4 66.5-25.4 91.9 0l429.6 429.5c25.4 25.4 25.4 66.5.1 91.9-12.3 12.3-28.8 19.1-46.1 19.1z"></path></svg></span>
-</h2>
-</a>
-
-
-
-<section class="o-mega-menu_content o-mega-menu_content-2 " aria-expanded="false" data-js-hook="behavior_flyout-menu_content">
-
-
-
-    <button class="o-mega-menu_content-alt-trigger o-mega-menu_content-2-alt-trigger " data-js-hook="behavior_flyout-menu_alt-trigger">
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 559.6 1200" class="cf-icon-svg"><path d="M494.5 1090.7c-17.3 0-33.8-6.8-46-19L19 642.1c-25.4-25.4-25.4-66.5 0-91.9l429.5-429.5c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L156.9 596.2l383.6 383.6c25.4 25.4 25.4 66.5.1 91.9-12.3 12.2-28.8 19-46.1 19z"></path></svg>
-        Back
-    </button>
-
-
-    <div class="o-mega-menu_content-wrapper o-mega-menu_content-2-wrapper ">
-
-
-        <div class="o-mega-menu_content-overview o-mega-menu_content-2-overview
-                    o-mega-menu_content-overview-heading o-mega-menu_content-2-overview-heading ">
-            <a class="o-mega-menu_content-overview-link o-mega-menu_content-2-overview-link
-                        " href="/consumer-tools/">
-                Consumer Tools Overview
-            </a>
-        </div>
-
-
-
-
-
-
-        <div class="o-mega-menu_content-lists o-mega-menu_content-2-lists ">
-
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-
-            <h3 class="h5
-                    o-mega-menu_group-heading
-                    " id="money-topics-menu" aria-label="Money Topics">
-                Money Topics
-            </h3>
-
-            <ul aria-labelledby="money-topics-menu">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/auto-loans/">
-
-
-
-
-        Auto Loans
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/bank-accounts/">
-
-
-
-
-        Bank Accounts &amp; Services
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/credit-cards/">
-
-
-
-
-        Credit Cards
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/credit-reports-and-scores/">
-
-
-
-
-        Credit Reports &amp; Scores
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/debt-collection/">
-
-
-
-
-        Debt Collection
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/fraud/">
-
-
-
-
-        Fraud &amp; Scams
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/money-transfers/">
-
-
-
-
-        Money Transfers
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-
-            <h3 class="h5
-                    o-mega-menu_group-heading
-                    o-mega-menu_group-heading__hidden" id="money-topics-continued-menu" aria-label="Money Topics continued">
-                Money Topics
-            </h3>
-
-            <ul aria-labelledby="money-topics-continued-menu">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/mortgages/">
-
-
-
-
-        Mortgages
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/payday-loans/">
-
-
-
-
-        Payday Loans
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/prepaid-cards/">
-
-
-
-
-        Prepaid Cards
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/reverse-mortgages/">
-
-
-
-
-        Reverse Mortgages
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="None">
-
-
-
-
-        Savings
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/student-loans/">
-
-
-
-
-        Student Loans
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-
-            <h3 class="h5
-                    o-mega-menu_group-heading
-                    " id="guides-menu" aria-label="Guides">
-                Guides
-            </h3>
-
-            <ul aria-labelledby="guides-menu">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/owning-a-home/">
-
-
-
-
-        Buying a House
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/getting-an-auto-loan/">
-
-
-
-
-        Getting an Auto Loan
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/managing-someone-elses-money/">
-
-
-
-
-        Managing Someone Else's Money
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/money-as-you-grow/">
-
-
-
-
-        Money as You Grow
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/military-financial-lifecycle/">
-
-
-
-
-        Navigating the Military Financial Lifecycle
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/paying-for-college/">
-
-
-
-
-        Paying for College
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/consumer-tools/retirement/">
-
-
-
-
-        Planning for Retirement
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-
-
-
-
-
-            <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-
-
-                <div class="o-mega-menu_content-list o-mega-menu_content-2-list
-                            o-mega-menu_content-list__featured o-mega-menu_content-2-list__featured ">
-                    <h3 class="h5 o-mega-menu_group-heading">
-                        Featured
-                    </h3>
-                    <ul aria-label="Featured">
-
-                        <li class="o-mega-menu_content-item o-mega-menu_content-2-item
-                                    o-mega-menu_content-item__has-icon o-mega-menu_content-2-item__has-icon ">
-                            <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/coronavirus/">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 969.7 1200" class="cf-icon-svg"><path d="M843.8 592.5l94.6-87.5c52.5-48.5 37.3-95.1-33.7-103.6l-128-15.2-129-15.3-81.5-176.6-27-58.5c-30-64.9-79-64.9-108.9 0l-27 58.5L322 371l-129.1 15.3-128 15.2c-71 8.4-86.1 55-33.6 103.6l47.3 43.7 47.3 43.7 95.4 88.3L196 808.2l-12.6 63.2-12.6 63.2c-13.9 70.1 25.7 98.9 88.1 64l225.9-126.4 225.9 126.4c62.4 34.9 102 6.1 88.1-64l-12.6-63.2-37.9-190.7 95.5-88.2z"></path></svg>
-                                <span><span class="a-link_text">Protect your finances during the Coronavirus Pandemic</span></span>
-                            </a>
-                        </li>
-
-                    </ul>
-                </div>
-
-
-
-                <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-                    <h3 class="h5 o-mega-menu_group-heading">
-                        Additional Resources
-                    </h3>
-                    <ul aria-label="Additional Resources">
-
-                        <li class="o-mega-menu_content-item o-mega-menu_content-2-item
-                                    o-mega-menu_content-item__has-icon o-mega-menu_content-2-item__has-icon ">
-                            <a class="a-link
-                                      o-mega-menu_content-link o-mega-menu_content-2-link
-                                      " href="/ask-cfpb/">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 609.9 1200" class="cf-icon-svg"><path d="M397.5 989.9H209.2c-8.3 0-15 6.7-15 15 0 4.6 2.1 9 5.8 11.8 60.8 47.3 146 47.3 206.8 0 6.5-5.1 7.7-14.5 2.6-21-2.9-3.7-7.2-5.8-11.9-5.8zM585.9 312.7C520.4 157.5 341.6 84.8 186.4 150.3 73.4 197.9-.1 308.7 0 431.4c0 106.3 52.5 177.8 90.9 229.9 17.3 23.5 36.9 50.2 36.9 63.1v65.4c0 33.1 26.9 60 60 60h234.3c33.1 0 60-26.9 60-60v-65.5c0-12.9 19.6-39.6 36.9-63.1 38.3-52.2 90.9-123.6 90.9-229.9.1-40.7-8.1-81.1-24-118.6zM422.1 724.5V790H187.8v-65.5c0-77.8-127.8-143.9-127.8-293 0-135.3 109.7-244.9 244.9-244.9s244.9 109.6 244.9 244.9c.1 149.1-127.7 215.2-127.7 293zM422.1 890.3H187.8c-16.6 0-30 13.4-30 30s13.4 30 30 30h234.3c16.6 0 30-13.4 30-30s-13.4-30-30-30z"></path></svg>
-                                <span><span class="a-link_text">Ask CFPB: Find answers to your money topic questions</span></span>
-                            </a>
-                        </li>
-
-                        <li class="o-mega-menu_content-item o-mega-menu_content-2-item
-                                    o-mega-menu_content-item__has-icon o-mega-menu_content-2-item__has-icon ">
-                            <a class="a-link
-                                      o-mega-menu_content-link o-mega-menu_content-2-link
-                                      " href="/your-story/">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 984.4 1200" class="cf-icon-svg"><path d="M400.7 674.5c-10.9-19.7-25.1-37.4-42-52.3-16.7-14.7-35.8-26.4-56.4-34.8-.5-.2-1.1-.4-1.6-.5-55-21.7-117-17-168.1 12.7 8.5-15.2 18.7-29.5 30.2-42.7 24.3-27.7 55.1-51.2 91.5-69.8 27.5-14.1 58.1-24.6 91-31.2 27.2-4.7 45.5-30.6 40.8-57.8s-30.6-45.5-57.8-40.8c-.9.2-1.8.3-2.7.5-41.9 8.4-81.2 22-116.9 40.2-47.7 24.4-88.5 55.7-121.2 92.9-32.7 37.4-57.2 81.2-71.8 128.7-15 47.7-19.2 99-12.6 152.3l.2 1.5.3 1.5c.2.9.4 1.8.5 2.8.2 1.2.5 2.4.7 3.6.2 1 .4 1.9.6 2.9l.9 4.2c.1.6.3 1.2.4 1.8.4 1.7.8 3.5 1.2 5.2l.1.4c7.2 28.5 17.8 53.4 31.7 74.5 17.4 26.4 37.7 47.5 60.9 63.3 22.8 15.7 48.6 26.6 75.7 32 26.9 5.5 54.5 6.2 81.6 2 26.3-3.8 51.5-13.1 74-27.2 21.8-13.6 40.7-31.2 55.8-52 15-20.8 26-44.3 32.4-69.2 6.7-25.8 7.7-52.6 3.1-78.8-3.5-23.1-11.1-45.4-22.5-65.9zM981.6 740.5c-6.9-45.8-29.6-87.7-64.4-118.4-16.7-14.7-35.8-26.4-56.4-34.8-.5-.2-1.1-.4-1.6-.5-55.1-21.7-117-17-168.2 12.7 8.5-15.2 18.7-29.5 30.2-42.7 24.3-27.7 55.1-51.2 91.5-69.8 27.5-14.1 58.1-24.6 91-31.2 26.9-6.2 43.7-33 37.5-59.9-5.9-25.9-31.1-42.6-57.2-38.1-41.9 8.4-81.2 22-116.9 40.2-47.7 24.4-88.5 55.7-121.2 92.9-32.7 37.4-57.2 81.2-71.9 128.7-14.9 47.7-19.2 99-12.6 152.3l.2 1.5.3 1.5c.2.9.4 1.8.5 2.8.2 1.2.5 2.4.7 3.6.2 1 .4 1.9.6 2.9l.9 4.2c.1.6.3 1.2.4 1.8.4 1.7.8 3.5 1.2 5.2l.1.4c7.2 28.5 17.8 53.4 31.7 74.5 17.4 26.4 37.7 47.5 60.9 63.3 22.8 15.7 48.6 26.6 75.7 32 26.9 5.5 54.5 6.2 81.6 2 26.3-3.8 51.5-13.1 74-27.2 21.8-13.6 40.7-31.2 55.8-52 15-20.8 26-44.3 32.4-69.2 6.8-25.6 7.8-52.5 3.2-78.7z"></path></svg>
-                                <span><span class="a-link_text">Tell Your Story</span></span>
-                            </a>
-                        </li>
-
-                        <li class="o-mega-menu_content-item o-mega-menu_content-2-item
-                                    o-mega-menu_content-item__has-icon o-mega-menu_content-2-item__has-icon ">
-                            <a class="a-link
-                                      o-mega-menu_content-link o-mega-menu_content-2-link
-                                      " href="/complaint/">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 825.1 1200" class="cf-icon-svg"><path d="M755.1 263.9H70c-38.6.1-69.9 31.4-70 70v459.5c.1 38.6 31.4 69.9 70 70h268.4l56.9 91.2c8.7 14 23 14 31.8 0l56.9-91.2h271.1c38.6-.1 69.9-31.4 70-70V333.9c-.1-38.7-31.4-69.9-70-70zm-375 132.5c0-17.9 14.6-32.5 32.5-32.5s32.5 14.6 32.5 32.5v204.2c0 17.9-14.6 32.5-32.5 32.5s-32.5-14.6-32.5-32.5V396.4zm32.5 364.4c-24.6 0-44.6-20-44.6-44.6s20-44.6 44.6-44.6 44.6 20 44.6 44.6c-.1 24.6-20 44.6-44.6 44.6z"></path></svg>
-                                <span><span class="a-link_text">Submit a Complaint</span></span>
-                            </a>
-                        </li>
-
-                    </ul>
-                </div>
-
-
-            </div>
-
-        </div>
-
-    </div>
-
-
-</section>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-1-item " data-js-hook="behavior_flyout-menu">
-
-    <a aria-haspopup="menu" class="o-mega-menu_content-link o-mega-menu_content-1-link o-mega-menu_content-link__has-children o-mega-menu_content-1-link__has-children " data-js-hook="behavior_flyout-menu_trigger" href="#">
-<h2>
-
-
-
-
-        Practitioner Resources
-
-
-        <span class="o-mega-menu_content-link-icon-post o-mega-menu_content-1-link-icon-post "><svg class="cf-icon-svg" viewBox="0 0 559.6 1200" xmlns="http://www.w3.org/2000/svg"><path d="M65.1 1090.2c-35.9 0-65-29-65.1-64.9 0-17.3 6.8-33.9 19.1-46.1l383.6-383.5L19.1 212.2c-25.1-25.6-24.8-66.8.9-92 25.3-24.8 65.8-24.8 91.1 0l429.5 429.5c25.4 25.4 25.4 66.5 0 91.9L111 1071.2c-12.1 12.2-28.7 19.1-45.9 19z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-closed o-mega-menu_content-1-link-icon-closed "><svg class="cf-icon-svg" viewBox="0 0 988.2 1200" xmlns="http://www.w3.org/2000/svg"><path d="M494.1 967.2c-17.3 0-33.8-6.8-46-19L18.6 518.6c-25.1-25.6-24.8-66.8.8-91.9 25.3-24.8 65.8-24.8 91.1 0l383.6 383.6 383.6-383.6c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L540.1 948.1c-12.2 12.2-28.7 19.1-46 19.1z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-open o-mega-menu_content-1-link-icon-open "><svg class="cf-icon-svg" viewBox="0 0 988.7 1200" xmlns="http://www.w3.org/2000/svg"><path d="M923.6 967.6c-17.3 0-33.8-6.8-46-19L494.1 565 110.5 948.5c-25.6 25.1-66.8 24.8-91.9-.8-24.8-25.3-24.8-65.8 0-91.1l429.5-429.5c25.4-25.4 66.5-25.4 91.9 0l429.6 429.5c25.4 25.4 25.4 66.5.1 91.9-12.3 12.3-28.8 19.1-46.1 19.1z"></path></svg></span>
-</h2>
-</a>
-
-
-
-<section class="o-mega-menu_content o-mega-menu_content-2 " aria-expanded="false" data-js-hook="behavior_flyout-menu_content">
-
-
-
-    <button class="o-mega-menu_content-alt-trigger o-mega-menu_content-2-alt-trigger " data-js-hook="behavior_flyout-menu_alt-trigger">
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 559.6 1200" class="cf-icon-svg"><path d="M494.5 1090.7c-17.3 0-33.8-6.8-46-19L19 642.1c-25.4-25.4-25.4-66.5 0-91.9l429.5-429.5c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L156.9 596.2l383.6 383.6c25.4 25.4 25.4 66.5.1 91.9-12.3 12.2-28.8 19-46.1 19z"></path></svg>
-        Back
-    </button>
-
-
-    <div class="o-mega-menu_content-wrapper o-mega-menu_content-2-wrapper ">
-
-
-
-
-
-
-
-        <div class="o-mega-menu_content-lists o-mega-menu_content-2-lists ">
-
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-
-            <h3 class="h5
-                    o-mega-menu_group-heading
-                    " id="populations-served-menu" aria-label="Populations Served">
-                Populations Served
-            </h3>
-
-            <ul aria-labelledby="populations-served-menu">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/adult-financial-education/">
-
-
-
-
-        Adult Financial Education
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/economically-vulnerable/">
-
-
-
-
-        Economically Vulnerable Consumers
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/resources-for-older-adults/">
-
-
-
-
-        Older Adults &amp; their Families
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-
-            <h3 class="h5
-                    o-mega-menu_group-heading
-                    o-mega-menu_group-heading__hidden" id="populations-served-continued-menu" aria-label="Populations Served continued">
-                Populations Served
-            </h3>
-
-            <ul aria-labelledby="populations-served-continued-menu">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/servicemembers/">
-
-
-
-
-        Servicemembers &amp; Veterans
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/students/">
-
-
-
-
-        Students &amp; Student Loan Borrowers
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/youth-financial-education/">
-
-
-
-
-        Youth Financial Education
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-
-            <h3 class="h5
-                    o-mega-menu_group-heading
-                    " id="programs-menu" aria-label="Programs">
-                Programs
-            </h3>
-
-            <ul aria-labelledby="programs-menu">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/employers/">
-
-
-
-
-        Employer Resources
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/financial-coaching/">
-
-
-
-
-        Financial Coaching
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/financial-well-being-resources/">
-
-
-
-
-        Financial Well-Being Resources
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/library-resources/">
-
-
-
-
-        Resources for Libraries
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/resources-for-tax-preparers/">
-
-
-
-
-        Resources for Tax Preparers
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/resources-youth-employment-programs/">
-
-
-
-
-        Resources for Youth Employment Programs
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/practitioner-resources/your-money-your-goals/">
-
-
-
-
-        Your Money, Your Goals
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-
-
-
-
-
-            <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-
-
-                <div class="o-mega-menu_content-list o-mega-menu_content-2-list
-                            o-mega-menu_content-list__featured o-mega-menu_content-2-list__featured ">
-                    <h3 class="h5 o-mega-menu_group-heading">
-                        Featured
-                    </h3>
-                    <ul aria-label="Featured">
-
-                        <li class="o-mega-menu_content-item o-mega-menu_content-2-item
-                                    o-mega-menu_content-item__has-icon o-mega-menu_content-2-item__has-icon ">
-                            <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/research-reports/financial-well-being-america/">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 969.7 1200" class="cf-icon-svg"><path d="M843.8 592.5l94.6-87.5c52.5-48.5 37.3-95.1-33.7-103.6l-128-15.2-129-15.3-81.5-176.6-27-58.5c-30-64.9-79-64.9-108.9 0l-27 58.5L322 371l-129.1 15.3-128 15.2c-71 8.4-86.1 55-33.6 103.6l47.3 43.7 47.3 43.7 95.4 88.3L196 808.2l-12.6 63.2-12.6 63.2c-13.9 70.1 25.7 98.9 88.1 64l225.9-126.4 225.9 126.4c62.4 34.9 102 6.1 88.1-64l-12.6-63.2-37.9-190.7 95.5-88.2z"></path></svg>
-                                <span><span class="a-link_text">Explore financial well-being survey results</span></span>
-                            </a>
-                        </li>
-
-                    </ul>
-                </div>
-
-
-
-                <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-                    <h3 class="h5 o-mega-menu_group-heading">
-                        Additional Resources
-                    </h3>
-                    <ul aria-label="Additional Resources">
-
-                        <li class="o-mega-menu_content-item o-mega-menu_content-2-item
-                                    o-mega-menu_content-item__has-icon o-mega-menu_content-2-item__has-icon ">
-                            <a class="a-link o-mega-menu_content-link o-mega-menu_content-2-link  a-link__icon" href="https://pueblo.gpo.gov/CFPBPubs/CFPBPubs.php">
-<svg class="cf-icon-svg" viewBox="0 0 651.7 1200" xmlns="http://www.w3.org/2000/svg"><path d="M30 161c-16.5 0-30 13.5-30 30v827.8c0 16.5 13.5 30 30 30h591.7c16.5 0 30-13.5 30-30V343.7L469 161H30zm389.6 60v134.8c0 19.9 16.3 36.2 36.2 36.2h135.9V988.8H60V221h359.6z"></path><path d="M123.8 768.6h394.8v50H123.8zM123.8 644h394.8v50H123.8zM123.8 519.5h394.8v50H123.8z"></path><circle cx="194" cy="382.3" r="60"></circle></svg>
-<span><span class="a-link_text">Order free brochures</span> <svg class="cf-icon-svg" viewBox="0 0 790.1 1200" xmlns="http://www.w3.org/2000/svg"><path d="M740.1 178.8H449.5c-27.6 0-50 22.4-50 50s22.4 50 50 50h169.9L518.2 379.9c-14.7-6.2-30.5-9.4-46.4-9.4H120c-66.2.1-119.9 53.8-120 120v351.8c.1 66.2 53.8 119.9 120 120h351.8c66.2-.1 119.9-53.8 120-120V490.6c0-12.6-2-25-5.8-37l104.1-104.1v169.9c0 27.6 22.4 50 50 50s50-22.4 50-50V228.8c0-27.7-22.4-50-50-50zM491.8 842.3c-.1 11-9 19.9-20 20H120c-11-.1-19.9-9-20-20V490.6c.1-11 9-19.9 20-20h307.5L267 631.1c-19.4 19.6-19.3 51.3.3 70.7 19.5 19.3 50.9 19.3 70.4 0l154-154 .1 294.5z"></path></svg></span>
-</a>
-                        </li>
-
-                    </ul>
-                </div>
-
-
-            </div>
-
-        </div>
-
-    </div>
-
-
-</section>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-1-item " data-js-hook="behavior_flyout-menu">
-
-    <a aria-haspopup="menu" class="o-mega-menu_content-link o-mega-menu_content-1-link o-mega-menu_content-link__has-children o-mega-menu_content-1-link__has-children " data-js-hook="behavior_flyout-menu_trigger" href="/data-research/">
-<h2>
-
-
-
-
-        Data &amp; Research
-
-
-        <span class="o-mega-menu_content-link-icon-post o-mega-menu_content-1-link-icon-post "><svg class="cf-icon-svg" viewBox="0 0 559.6 1200" xmlns="http://www.w3.org/2000/svg"><path d="M65.1 1090.2c-35.9 0-65-29-65.1-64.9 0-17.3 6.8-33.9 19.1-46.1l383.6-383.5L19.1 212.2c-25.1-25.6-24.8-66.8.9-92 25.3-24.8 65.8-24.8 91.1 0l429.5 429.5c25.4 25.4 25.4 66.5 0 91.9L111 1071.2c-12.1 12.2-28.7 19.1-45.9 19z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-closed o-mega-menu_content-1-link-icon-closed "><svg class="cf-icon-svg" viewBox="0 0 988.2 1200" xmlns="http://www.w3.org/2000/svg"><path d="M494.1 967.2c-17.3 0-33.8-6.8-46-19L18.6 518.6c-25.1-25.6-24.8-66.8.8-91.9 25.3-24.8 65.8-24.8 91.1 0l383.6 383.6 383.6-383.6c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L540.1 948.1c-12.2 12.2-28.7 19.1-46 19.1z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-open o-mega-menu_content-1-link-icon-open "><svg class="cf-icon-svg" viewBox="0 0 988.7 1200" xmlns="http://www.w3.org/2000/svg"><path d="M923.6 967.6c-17.3 0-33.8-6.8-46-19L494.1 565 110.5 948.5c-25.6 25.1-66.8 24.8-91.9-.8-24.8-25.3-24.8-65.8 0-91.1l429.5-429.5c25.4-25.4 66.5-25.4 91.9 0l429.6 429.5c25.4 25.4 25.4 66.5.1 91.9-12.3 12.3-28.8 19.1-46.1 19.1z"></path></svg></span>
-</h2>
-</a>
-
-
-
-<section class="o-mega-menu_content o-mega-menu_content-2 " aria-expanded="false" data-js-hook="behavior_flyout-menu_content">
-
-
-
-    <button class="o-mega-menu_content-alt-trigger o-mega-menu_content-2-alt-trigger " data-js-hook="behavior_flyout-menu_alt-trigger">
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 559.6 1200" class="cf-icon-svg"><path d="M494.5 1090.7c-17.3 0-33.8-6.8-46-19L19 642.1c-25.4-25.4-25.4-66.5 0-91.9l429.5-429.5c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L156.9 596.2l383.6 383.6c25.4 25.4 25.4 66.5.1 91.9-12.3 12.2-28.8 19-46.1 19z"></path></svg>
-        Back
-    </button>
-
-
-    <div class="o-mega-menu_content-wrapper o-mega-menu_content-2-wrapper ">
-
-
-        <div class="o-mega-menu_content-overview o-mega-menu_content-2-overview
-                    o-mega-menu_content-overview-heading o-mega-menu_content-2-overview-heading ">
-            <a class="o-mega-menu_content-overview-link o-mega-menu_content-2-overview-link
-                        " href="/data-research/">
-                Data &amp; Research Overview
-            </a>
-        </div>
-
-
-
-
-
-
-        <div class="o-mega-menu_content-lists o-mega-menu_content-2-lists ">
-
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="Data &amp; Research">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/research-reports/">
-
-
-
-
-        Research &amp; Reports
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/consumer-complaints/">
-
-
-
-
-        Consumer Complaint Database
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/hmda/">
-
-
-
-
-        Mortgage Database (HMDA)
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="Data &amp; Research">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/consumer-credit-trends/">
-
-
-
-
-        Consumer Credit Trends
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/mortgage-performance-trends/">
-
-
-
-
-        Mortgage Performance Trends
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/financial-well-being-survey-data/">
-
-
-
-
-        Financial Well-Being Survey
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="Data &amp; Research">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/credit-card-data/">
-
-
-
-
-        Credit Card Surveys &amp; Agreements
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/prepaid-accounts/">
-
-
-
-
-        Prepaid Account Agreements
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/cfpb-research-conference/">
-
-
-
-
-        CFPB Research Conference
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/cfpb-researchers/">
-
-
-
-
-        CFPB Researchers
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-
-
-
-
-
-            <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-
-
-                <div class="o-mega-menu_content-list o-mega-menu_content-2-list
-                            o-mega-menu_content-list__featured o-mega-menu_content-2-list__featured ">
-                    <h3 class="h5 o-mega-menu_group-heading">
-                        Featured
-                    </h3>
-                    <ul aria-label="Featured">
-
-                        <li class="o-mega-menu_content-item o-mega-menu_content-2-item
-                                    o-mega-menu_content-item__has-icon o-mega-menu_content-2-item__has-icon ">
-                            <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/data-research/financial-well-being-survey-data/">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 969.7 1200" class="cf-icon-svg"><path d="M843.8 592.5l94.6-87.5c52.5-48.5 37.3-95.1-33.7-103.6l-128-15.2-129-15.3-81.5-176.6-27-58.5c-30-64.9-79-64.9-108.9 0l-27 58.5L322 371l-129.1 15.3-128 15.2c-71 8.4-86.1 55-33.6 103.6l47.3 43.7 47.3 43.7 95.4 88.3L196 808.2l-12.6 63.2-12.6 63.2c-13.9 70.1 25.7 98.9 88.1 64l225.9-126.4 225.9 126.4c62.4 34.9 102 6.1 88.1-64l-12.6-63.2-37.9-190.7 95.5-88.2z"></path></svg>
-                                <span><span class="a-link_text">Help advance financial well-being</span></span>
-                            </a>
-                        </li>
-
-                    </ul>
-                </div>
-
-
-
-
-            </div>
-
-        </div>
-
-    </div>
-
-
-</section>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-1-item " data-js-hook="behavior_flyout-menu">
-
-    <a aria-haspopup="menu" class="o-mega-menu_content-link o-mega-menu_content-1-link o-mega-menu_content-link__has-children o-mega-menu_content-1-link__has-children " data-js-hook="behavior_flyout-menu_trigger" href="/policy-compliance/">
-<h2>
-
-
-
-
-        Policy &amp; Compliance
-
-
-        <span class="o-mega-menu_content-link-icon-post o-mega-menu_content-1-link-icon-post "><svg class="cf-icon-svg" viewBox="0 0 559.6 1200" xmlns="http://www.w3.org/2000/svg"><path d="M65.1 1090.2c-35.9 0-65-29-65.1-64.9 0-17.3 6.8-33.9 19.1-46.1l383.6-383.5L19.1 212.2c-25.1-25.6-24.8-66.8.9-92 25.3-24.8 65.8-24.8 91.1 0l429.5 429.5c25.4 25.4 25.4 66.5 0 91.9L111 1071.2c-12.1 12.2-28.7 19.1-45.9 19z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-closed o-mega-menu_content-1-link-icon-closed "><svg class="cf-icon-svg" viewBox="0 0 988.2 1200" xmlns="http://www.w3.org/2000/svg"><path d="M494.1 967.2c-17.3 0-33.8-6.8-46-19L18.6 518.6c-25.1-25.6-24.8-66.8.8-91.9 25.3-24.8 65.8-24.8 91.1 0l383.6 383.6 383.6-383.6c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L540.1 948.1c-12.2 12.2-28.7 19.1-46 19.1z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-open o-mega-menu_content-1-link-icon-open "><svg class="cf-icon-svg" viewBox="0 0 988.7 1200" xmlns="http://www.w3.org/2000/svg"><path d="M923.6 967.6c-17.3 0-33.8-6.8-46-19L494.1 565 110.5 948.5c-25.6 25.1-66.8 24.8-91.9-.8-24.8-25.3-24.8-65.8 0-91.1l429.5-429.5c25.4-25.4 66.5-25.4 91.9 0l429.6 429.5c25.4 25.4 25.4 66.5.1 91.9-12.3 12.3-28.8 19.1-46.1 19.1z"></path></svg></span>
-</h2>
-</a>
-
-
-
-<section class="o-mega-menu_content o-mega-menu_content-2 " aria-expanded="false" data-js-hook="behavior_flyout-menu_content">
-
-
-
-    <button class="o-mega-menu_content-alt-trigger o-mega-menu_content-2-alt-trigger " data-js-hook="behavior_flyout-menu_alt-trigger">
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 559.6 1200" class="cf-icon-svg"><path d="M494.5 1090.7c-17.3 0-33.8-6.8-46-19L19 642.1c-25.4-25.4-25.4-66.5 0-91.9l429.5-429.5c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L156.9 596.2l383.6 383.6c25.4 25.4 25.4 66.5.1 91.9-12.3 12.2-28.8 19-46.1 19z"></path></svg>
-        Back
-    </button>
-
-
-    <div class="o-mega-menu_content-wrapper o-mega-menu_content-2-wrapper ">
-
-
-        <div class="o-mega-menu_content-overview o-mega-menu_content-2-overview
-                    o-mega-menu_content-overview-heading o-mega-menu_content-2-overview-heading ">
-            <a class="o-mega-menu_content-overview-link o-mega-menu_content-2-overview-link
-                        " href="/policy-compliance/">
-                Policy &amp; Compliance Overview
-            </a>
-        </div>
-
-
-
-
-
-
-        <div class="o-mega-menu_content-lists o-mega-menu_content-2-lists ">
-
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="Policy &amp; Compliance">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/policy-compliance/rulemaking/">
-
-
-
-
-        Rulemaking
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="Policy &amp; Compliance">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/policy-compliance/guidance/">
-
-
-
-
-        Compliance &amp; Guidance
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/policy-compliance/enforcement/">
-
-
-
-
-        Enforcement
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="Policy &amp; Compliance">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/policy-compliance/notice-opportunities-comment/">
-
-
-
-
-        Notices &amp; Opportunities to Comment
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/policy-compliance/amicus/">
-
-
-
-
-        Amicus Program
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-
-
-
-
-
-            <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-
-
-                <div class="o-mega-menu_content-list o-mega-menu_content-2-list
-                            o-mega-menu_content-list__featured o-mega-menu_content-2-list__featured ">
-                    <h3 class="h5 o-mega-menu_group-heading">
-                        Featured
-                    </h3>
-                    <ul aria-label="Featured">
-
-                        <li class="o-mega-menu_content-item o-mega-menu_content-2-item
-                                    o-mega-menu_content-item__has-icon o-mega-menu_content-2-item__has-icon ">
-                            <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/policy-compliance/guidance/">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 969.7 1200" class="cf-icon-svg"><path d="M843.8 592.5l94.6-87.5c52.5-48.5 37.3-95.1-33.7-103.6l-128-15.2-129-15.3-81.5-176.6-27-58.5c-30-64.9-79-64.9-108.9 0l-27 58.5L322 371l-129.1 15.3-128 15.2c-71 8.4-86.1 55-33.6 103.6l47.3 43.7 47.3 43.7 95.4 88.3L196 808.2l-12.6 63.2-12.6 63.2c-13.9 70.1 25.7 98.9 88.1 64l225.9-126.4 225.9 126.4c62.4 34.9 102 6.1 88.1-64l-12.6-63.2-37.9-190.7 95.5-88.2z"></path></svg>
-                                <span><span class="a-link_text">Resources to help you comply</span></span>
-                            </a>
-                        </li>
-
-                    </ul>
-                </div>
-
-
-
-
-            </div>
-
-        </div>
-
-    </div>
-
-
-</section>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-1-item " data-js-hook="behavior_flyout-menu">
-
-    <a aria-haspopup="menu" class="o-mega-menu_content-link o-mega-menu_content-1-link o-mega-menu_content-link__has-children o-mega-menu_content-1-link__has-children " data-js-hook="behavior_flyout-menu_trigger" href="/about-us/">
-<h2>
-
-
-
-
-        About Us
-
-
-        <span class="o-mega-menu_content-link-icon-post o-mega-menu_content-1-link-icon-post "><svg class="cf-icon-svg" viewBox="0 0 559.6 1200" xmlns="http://www.w3.org/2000/svg"><path d="M65.1 1090.2c-35.9 0-65-29-65.1-64.9 0-17.3 6.8-33.9 19.1-46.1l383.6-383.5L19.1 212.2c-25.1-25.6-24.8-66.8.9-92 25.3-24.8 65.8-24.8 91.1 0l429.5 429.5c25.4 25.4 25.4 66.5 0 91.9L111 1071.2c-12.1 12.2-28.7 19.1-45.9 19z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-closed o-mega-menu_content-1-link-icon-closed "><svg class="cf-icon-svg" viewBox="0 0 988.2 1200" xmlns="http://www.w3.org/2000/svg"><path d="M494.1 967.2c-17.3 0-33.8-6.8-46-19L18.6 518.6c-25.1-25.6-24.8-66.8.8-91.9 25.3-24.8 65.8-24.8 91.1 0l383.6 383.6 383.6-383.6c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L540.1 948.1c-12.2 12.2-28.7 19.1-46 19.1z"></path></svg></span>
-<span class="o-mega-menu_content-link-icon-open o-mega-menu_content-1-link-icon-open "><svg class="cf-icon-svg" viewBox="0 0 988.7 1200" xmlns="http://www.w3.org/2000/svg"><path d="M923.6 967.6c-17.3 0-33.8-6.8-46-19L494.1 565 110.5 948.5c-25.6 25.1-66.8 24.8-91.9-.8-24.8-25.3-24.8-65.8 0-91.1l429.5-429.5c25.4-25.4 66.5-25.4 91.9 0l429.6 429.5c25.4 25.4 25.4 66.5.1 91.9-12.3 12.3-28.8 19.1-46.1 19.1z"></path></svg></span>
-</h2>
-</a>
-
-
-
-<section class="o-mega-menu_content o-mega-menu_content-2 " aria-expanded="false" data-js-hook="behavior_flyout-menu_content">
-
-
-
-    <button class="o-mega-menu_content-alt-trigger o-mega-menu_content-2-alt-trigger " data-js-hook="behavior_flyout-menu_alt-trigger">
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 559.6 1200" class="cf-icon-svg"><path d="M494.5 1090.7c-17.3 0-33.8-6.8-46-19L19 642.1c-25.4-25.4-25.4-66.5 0-91.9l429.5-429.5c25.6-25.1 66.8-24.8 91.9.8 24.8 25.3 24.8 65.8 0 91.1L156.9 596.2l383.6 383.6c25.4 25.4 25.4 66.5.1 91.9-12.3 12.2-28.8 19-46.1 19z"></path></svg>
-        Back
-    </button>
-
-
-    <div class="o-mega-menu_content-wrapper o-mega-menu_content-2-wrapper ">
-
-
-        <div class="o-mega-menu_content-overview o-mega-menu_content-2-overview
-                    o-mega-menu_content-overview-heading o-mega-menu_content-2-overview-heading ">
-            <a class="o-mega-menu_content-overview-link o-mega-menu_content-2-overview-link
-                        " href="/about-us/">
-                About Us Overview
-            </a>
-        </div>
-
-
-
-
-
-
-        <div class="o-mega-menu_content-lists o-mega-menu_content-2-lists ">
-
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="About Us">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/the-bureau/">
-
-
-
-
-        The Bureau
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/budget-strategy/">
-
-
-
-
-        Budget &amp; Strategy
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/payments-harmed-consumers/">
-
-
-
-
-        Payments to Harmed Consumers
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/diversity-and-inclusion/">
-
-
-
-
-        Diversity &amp; Inclusion
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="About Us">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/activity-log/">
-
-
-
-
-        Recent Updates
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/blog/">
-
-
-
-
-        Blog
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/newsroom/">
-
-
-
-
-        Newsroom
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/events/">
-
-
-
-
-        Events
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="About Us">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/careers/">
-
-
-
-
-        Careers
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/doing-business-with-us/">
-
-
-
-
-        Doing Business With Us
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/advisory-committees/">
-
-
-
-
-        Advisory Committees
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/innovation/">
-
-
-
-
-        Innovation
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-    <div class="o-mega-menu_content-list-group o-mega-menu_content-2-list-group ">
-        <div class="o-mega-menu_content-list o-mega-menu_content-2-list ">
-
-
-            <ul aria-label="About Us">
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/contact-us/">
-
-
-
-
-        Contact Us
-
-
-
-
-    </a>
-
-
-</li>
-
-
-
-
-<li class="o-mega-menu_content-item o-mega-menu_content-2-item ">
-
-    <a class="o-mega-menu_content-link o-mega-menu_content-2-link " href="/about-us/taskforce-federal-consumer-financial-law/">
-
-
-
-
-        Taskforce on Federal Consumer Financial Law
-
-
-
-
-    </a>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-
-
-
-
-        </div>
-
-    </div>
-
-
-</section>
-
-
-</li>
-
-            </ul>
-
-        </div>
-    </div>
-
-
-
-
-
-        </div>
-
-    </div>
-
-
-
-
-
-    <div class="m-global-eyebrow
-                m-global-eyebrow__list">
-        <div class="wrapper
-                    ">
-
-<div class="a-tagline">
-    An official website of the United States government
-</div>
-
-            <div class="m-global-eyebrow_actions">
-                <ul class="m-list
-                           m-list__horizontal
-                           m-global-eyebrow_languages">
-                    <li class="m-list_item">
-
-                        <a href="/es/" hreflang="es" lang="es">
-                            Español
-                        </a>
-
-                    </li>
-
-                    <li class="m-list_item">
-                        <a href="/language/zh/" hreflang="zh" lang="zh">
-                            中文
-                        </a>
-                    </li>
-                    <li class="m-list_item">
-                        <a href="/language/vi/" hreflang="vi" lang="vi">
-                            Tiếng Việt
-                        </a>
-                    </li>
-                    <li class="m-list_item">
-                        <a href="/language/ko/" hreflang="ko" lang="ko">
-                            한국어
-                        </a>
-                    </li>
-                    <li class="m-list_item">
-                        <a href="/language/tl/" hreflang="tl" lang="tl">
-                            Tagalog
-                        </a>
-                    </li>
-                    <li class="m-list_item">
-                        <a href="/language/ru/" hreflang="ru" lang="ru">
-                            Pусский
-                        </a>
-                    </li>
-                    <li class="m-list_item">
-                        <a href="/language/ar/" hreflang="ar" lang="ar">
-                            العربية
-                        </a>
-                    </li>
-                    <li class="m-list_item">
-                        <a href="/language/ht/" hreflang="ht" lang="ht">
-                            Kreyòl Ayisyen
-                        </a>
-                    </li>
-
-                </ul>
-                <span class="m-global-eyebrow_phone">
-                    <a href="tel:+1-855-411-2372">(855) 411-2372</a>
-                </span>
-            </div>
-        </div>
-    </div>
-
-
-</section>
-
-</nav>
-`;
+import HTML_SNIPPET from '../../mocks/megaMenuSnippet';
 
 describe( 'MegaMenu', () => {
   let navElem;
@@ -2096,7 +9,7 @@ describe( 'MegaMenu', () => {
 
   beforeEach( () => {
     document.body.innerHTML = HTML_SNIPPET;
-    navElem = document.querySelector( `.${ BASE_CLASS }` );
+    navElem = document.querySelector( '.o-mega-menu' );
     megaMenu = new MegaMenu( navElem );
     thisMegaMenu = megaMenu.init();
   } );
